edit delete note api

+++ module/notes/business/delete_use_case.go
@@ -2,6 +2,7 @@ package business
 
 import (
 	"errors"
+	"gin-project/common"
 	"gin-project/module/notes/model"
 )
 
@@ -11,14 +12,16 @@ type DeleteUseCase interface {
 }
 
 type DeleteUseCaseImpl struct {
-	usecase DeleteUseCase
+	usecase   DeleteUseCase
+	requester common.Requester
 }
 
-func NewInstanceDeleteUseCase(useCase DeleteUseCase) *DeleteUseCaseImpl {
-	return &DeleteUseCaseImpl{useCase}
+func NewInstanceDeleteUseCase(useCase DeleteUseCase, requester common.Requester) *DeleteUseCaseImpl {
+	return &DeleteUseCaseImpl{useCase, requester}
 }
 
 func (impl *DeleteUseCaseImpl) DeleteNote(noteId int) error {
+
 	//find note
 	note, err := impl.usecase.FindNote(noteId)
 
@@ -29,6 +32,15 @@ func (impl *DeleteUseCaseImpl) DeleteNote(noteId int) error {
 	if note.Status == 0 {
 		return errors.New("note is deleted")
 	}
+
+	//check permission
+	isAuthor := impl.requester.GetUserId() == note.UserId
+	isAdmin := impl.requester.GetRole() == common.RoleAdmin
+
+	if !isAuthor && !isAdmin {
+		return common.ErrNoPermission(nil)
+	}
+
 	//end delete
 	if err := impl.usecase.DeleteNote(noteId); err != nil {
 		return errors.New("delete error is deleted")
